Refaktorert hvordan lagring av inntektsopplysnigner kalles
